requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-ollama
sentence_transformers
langchain_huggingface
tqdm
hf_xet
faiss-cpu
langchain_chroma
streamlit
langsmith
langchain_groq
fastapi
uvicorn
langserve[all]
langchain_core
duckdb
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
unstructured
whisper
pytube
numexpr
huggingface_hub
astrapy
cassio
datasets
tiktoken
PyPDF2
pdfminer
pdfminer.six
gradio
ddgs
altair
boto3
awscli
response
sagemaker
langchain_nvidia_ai_endpoints
crewai
crewai_tools
pinecone
pinecone-text
pinecone-notebooks
neo4j
langchain_neo4j
lamini
langchain_experimental
openpipe
langgraph
graphviz
pyppeteer
pydantic
langgraph-cli[inmem]
